refix

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -5,11 +5,10 @@
  * @s: char type string
  * Return: integer converted
  */
-
 int _atoi(char *s)
 {
 	int i;
-	long res = 0;
+	int res = 0;
 	int sig = 1;
 	int brk = 0;
